ai_engine/requirements.txt
fastapi
uvicorn
opencv-python-headless
ultralytics
openai
python-dotenv
numpy
requests
pydantic<2.0
sse-starlette
python-multipart
#py -3.11 -m uvicorn main:app --host 0.0.0.0 --port 8000